LayoutTests/platform/glib/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/no-referrer-from-meta-url-censored-expected.txt



FAIL The url of a document with no-referrer referrer meta tag is censored in NavigationHistoryEntry assert_equals: expected (object) null but got (string) "http://web-platform.test:8800/navigation-api/navigation-history-entry/resources/no-referrer-meta.html"
